requirements.txt
# Browser automation
zendriver>=0.5.0

# HTTP/async
aiohttp>=3.9.0
aiofiles>=24.1.0
httpx>=0.27.0

# AI/ML
openai-whisper>=20231117
anthropic>=0.40.0

# Image/Video processing
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0

# Database/Storage
supabase>=2.0.0
google-cloud-storage>=2.14.0
python-dotenv>=1.0.0
